+++ github_repo_changes.rb
@@ -5,12 +5,11 @@ require 'yaml'
 class GithubRepoChanges
   attr_reader :version_regex
   def initialize(config_file: 'config.yml',
-                 repo: nil,
-                 version_regex: /\d*\.\d*.\d*-\d*/)
+                 repo: nil)
     @config = YAML.load_file(config_file)
     @repo = repo
     @tags_filter = @config['tags_filter']
-    @version_regex = version_regex
+    @version_regex = /#{@config['version_regexp']}/
     Octokit.configure do |c|
       c.login = @config['github_user']
       c.password = @config['github_user_password']