+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Ability to get changes from specific version: ask `/get_changes 3.8.0-208`
 * Move setting tag filter to config file
 * More string checking for filter tags, after tags filter should be version numbers
+* Move regexp for version to config file
 
 ## 0.0.2
 ### New Features